backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
PyMySQL==1.0.2
python-dotenv==0.19.0